refactor: extract kv-dtype kernel requirement table to kernel_requirements.rs
init_kernel_dispatch.rs crossed the 500-LoC cap (510) after gaining the
startup requirement table. Pure code move per the Atlas split idiom:
required_optional_kernels_for_dtype + validate_required_kernels and
their tests now live in the sibling kernel_requirements.rs (214 LoC);
the dispatch table file is back to 321.
